// libs/routes/src/enums/route-path.ts
export enum RoutePath {
  Root = '/',
  BetaStaffUsers = '/beta',
  Dashboard = '/dashboard',
  EmailMessages = '/email_messages',
  RoleEmailMessages = '/:entityType/:entityId(\\d+):_?/email_messages',
  EntityPerformedActions = '/:entityType/:entityId(\\d+):_?/performed_actions',
  EntityGlobalPerformedActions = '/:entityType/performed_actions',
  EmailMessage = '/email_messages/:id(\\d+):_?',
  RoleEmailMessage = '/:entityType/:entityId(\\d+):_?/email_messages/:id(\\d+):_?',
  JobStation = '/job_station',
  CallRequests = '/callback_requests',
  CallRequest = '/callback_requests/:id(\\d+):_?',
  MeetingProfile = '/meetings/:id(\\d+):_?',
  Meetings = '/meetings/:category(unresolved|future)?',
  TalentScreeningSpecialistsTalents = '/tss_talent_list',
  Tasks = '/tasks',
  TasksByClients = '/tasks/companies_list',
  Playbook = '/playbooks/:identifier(\\w+):_?',
  ApplicantCompanies = '/applicants/companies',
  Clients = '/clients',
  ClientProfile = '/clients/:id(\\d+):_?',
  CompanyRepresentative = '/company_representatives/:id(\\d+):_?',
  Opportunity = '/opportunities/:id(\\d+):_?',
  CreateCompanyRepresentative = '/clients/:clientId(\\d+):_?/company_representatives/create',
  ClientHierarchy = '/clients/:id(\\d+):_?/hierarchy',
  EditCompanyRepresentative = '/company_representatives/:id(\\d+):_?/edit',
  BillingInvoices = '/invoices',
  BillingInvoice = '/invoices/:id(\\d+):_?',
  BillingPaymentReconciliation = '/invoices/reconciliation/',
  BillingPayments = '/payments',
  BillingPayment = '/payments/:id(\\d+):_?',
  BillingPurchaseOrders = '/purchase_orders',
  BillingPurchaseOrder = '/purchase_orders/:id(\\d+):_?',
  BillingPurchaseOrderLineLegacy = '/purchase_order_lines/:id(\\d+):_?',
  BillingPurchaseOrderLine = '/purchase_orders/:poId(\\d+):_?/lines/:id(\\d+):_?',
  BillingMemorandums = '/memos',
  BillingExpectedCommissions = '/expected_commissions/search',
  BillingPaymentGroups = '/payment_groups',
  BillingPaymentGroup = '/payment_groups/:id(\\d+):_?',
  ReceivedPayments = '/my_payments',
  MyExpectedCommissions = '/expected_commissions',
  SourcedTalents = '/sourced_talents',
  TalentProfile = '/talents/:id(\\d+):_?',
  TalentCreate = '/talents/new',
  TalentUpdate = '/talents/:id(\\d+):_?/edit',
  Talents = '/talents',
  GigCandidatesSearch = '/gig_candidates',
  TalentsForCoaching = '/talent_coaching',
  JobApplication = '/job_applications/:id(\\d+):_?',
  Jobs = '/jobs',
  Job = '/jobs/:id(\\d+):_?',
  JobBilling = '/jobs/:id(\\d+):_?/billing',
  JobCreate = '/jobs/new',
  JobUpdate = '/jobs/:id(\\d+):_?/edit',
  SourcingRequestList = '/sourcing_requests',
  AddSourcingRequest = '/sourcing_requests/create',
  EditSourcingRequest = '/sourcing_requests/:id(\\d+):_?/update',
  TalentInfractions = '/talent_infractions',
  Requests = '/toptal-publications',
  Request = '/toptal-publications/:id(\\w+):_?',
  PeerRequests = '/ask-a-peer',
  PeerRequest = '/ask-a-peer/:id(\\w+):_?',
  Skills = '/skills',
  Engagement = '/engagements/:id(\\d+):_?',
  CandidateSending = '/engagements/new',
  Calls = '/calls',
  UnfilledCalls = '/calls/unfilled',
  Quizzes = '/quizzes',
  CommunityLeaders = '/community_leaders',
  CommunityLeadersSort = '/community_leaders/active/sort',
  CommunityLeader = '/community_leaders/:id(\\w+):_?',
  CommunityEvents = '/community_events',
  CommunityEvent = '/community_events/:id(\\w+):_?',
  CommunityEventTags = '/community_event_tags',
  VerticalSettings = '/vertical_settings',
  AddVertical = '/vertical_settings/create',
  RateChangeRequests = '/rate_change_requests',
  Staff = '/staff',
  StaffProfile = '/staff/:id(\\d+):_?',
  Teams = '/teams',
  TeamProfile = '/teams/:id(\\d+):_?',
  EmailTemplates = '/email_templates',
  EmailTemplateCreate = '/email_templates/new',
  EmailTemplateUpdate = '/email_templates/:id(\\d+):_?/edit',
  EmailTemplateHelp = '/email_templates/help',
  TalentApplicants = '/applicants/talents',
  EmailTemplate = '/email_templates/:id(\\d+):_?',
  PageNotFound = '404'
}